% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.alphadiv.R
\name{analyze.alphadiv}
\alias{analyze.alphadiv}
\title{Create a alpha diversity profile of the retrieved data}
\usage{
analyze.alphadiv(
  bin.comm,
  plot.curve = FALSE,
  curve.index = NULL,
  curve.xval = NULL,
  preston.res = FALSE,
  pres.plot.y.label = NULL
)
}
\arguments{
\item{bin.comm}{the site X species like output from the \code{gen.comm.mat} function.Default is FALSE}

\item{plot.curve}{A logical value specifying whether a accummulation curve should be plotted}

\item{curve.index}{A character value specifying which index should be used for the \code{curve.index} argument. Default is NULL.}

\item{curve.xval}{A character value specifying whether sample or individuals should be used against the \code{curve.index}.Default is NULL}

\item{preston.res}{A logical value specifying if preston results should be generated. Default is FALSE.}

\item{pres.plot.y.label}{A character value specifying the taxonomic category which was used  to generate the matrix using the \code{gen.comm.mat} function}
}
\value{
A data frame containing all the information related to the processids/sampleids and the filters applied (if/any)
richness = A richness profile matrix
output$Shannon_div = shannon diversity values for the given sites/grids
output$richness_plot = A ggplot2 visualization of the richness curve
output$preston.res = a preston plot numerical data output
output$preston.plot = a ggplot2 visualization of the preston.plot
}
\description{
Function analyzes the output from the \code{gen.comm.mat} to provide a richness and alpha diversity profile of the downloaded data
}
\details{
analyze.alphadiv generates a alpha diversity profile along with richness estimations from the \code{gen.comm.mat} output. The estimations are based on BIN counts or presence absence data at the taxonomic level specified by the user in the \code{gen.comm.mat}function. The function also offers Shannon diversity values and Preston plots and the associated numerical results. Please note that the user should consider using some of the results (like Shannon/Preston) based on the input data (true abundances vs counts vs incidences). A richness profile is created using the 'alpha.accum' function from the BAT package while the preston and shannon diversity results are obtained using the 'prestondistr' and 'diversity' functions from vegan. Preston plots are created using the data from the 'prestondistr' results in 'ggplot2'
}
