% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.betadiv.R
\name{analyze.betadiv}
\alias{analyze.betadiv}
\title{Create a beta diversity profile of the retrieved data}
\usage{
analyze.betadiv(
  bin.comm,
  index,
  pre.abs = FALSE,
  heatmap = FALSE,
  component,
  grids = FALSE,
  grids.df = NULL
)
}
\arguments{
\item{bin.comm}{the site X species like output from the \code{gen.comm.mat} function.Default is FALSE}

\item{index}{A character vector specifying the type of beta diversity index ('jaccard' or 'sorenson' currently available)}

\item{pre.abs}{A logical value specifying if input data is presence-absence. Default is FALSE.}

\item{heatmap}{A logical value specifying whether a heatmap of the beta diversity values should be plotted.Default is FALSE.}

\item{component}{A character value specifying which beta diversity component should be used for the heatmap. Default is NULL.}

\item{grids}{A logical value specifying Whether the community matrix is generated using grids. Default is FALSE.}

\item{grids.df}{If grids = TRUE, a 'sf' grid data frame generated along with the community matrix using the\code{gen.comm.mat} function.Default is NULL}
}
\value{
A list containing:
\itemize{
\item output$total.beta = beta.total
\item output$replace = beta.replace (replacement)
\item output$richnessd = beta.richnessd (richness difference)
\item output$heatmap.viz = heatmap_final
}
}
\description{
Function analyzes the output from the \code{gen.comm.mat} to provide a beta diversity (and its partitions) profile of the downloaded data
}
\details{
The function calculates a sorenson or jaccard beta dissimilarity matrix using the \code{gen.comm.mat} output. It also generates matrices of 'species replacement' and 'richness difference' components of the total beta diversity. The values are calculated using \code{\link[BAT:beta]{BAT::beta()}} function which partitions the data using the Podani approach. A corresponding 'heatmap' can also be obtained when \code{heatmap}=TRUE. In case of grid based heatmaps, grids are arranged on the heatmap based on their centroid distances (i.e. nearest grids are placed closest). For site categories, the heatmap labels are arranged alphabetically. Grid based heatmaps can only be generated when \code{grids} = TRUE and a \link{sf} 'grid.df' which is generated from the \code{gen.comm.mat} function is provided to the function.
}
