% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.fetch.filters.R
\name{bold.connectr.filters}
\alias{bold.connectr.filters}
\title{Filters for specific parameters to customize the search for private data}
\usage{
bold.connectr.filters(
  bold.df,
  taxon.name = NULL,
  location.name = NULL,
  latitude = NULL,
  longitude = NULL,
  shapefile = NULL,
  institutes = NULL,
  identified.by = NULL,
  seq.source = NULL,
  marker = NULL,
  collection.period = NULL,
  basecount = NULL,
  altitude = NULL,
  depth = NULL
)
}
\arguments{
\item{bold.df}{the output from the connectr functions}

\item{taxon.name}{A single character string or a vector of taxonomic names at any hierarchical level. Default value is NULL}

\item{location.name}{A single character string or a vector of  country/province/state/region/sector/site names/codes. Default value is NULL.}

\item{latitude}{A number or a vector of two numbers indicating the latitudinal range. Values separated by a comma. Default value is NULL.}

\item{longitude}{A number or a vector of two numbers indicating the longitudinal range. Values separated by a comma. Default value is NULL.}

\item{shapefile}{A file path pointing to a shapefile or name of the shapefile imported in the session. Default value is NULL.}

\item{institutes}{A character string or a vector specifying names of institutes. Default value is NULL.}

\item{identified.by}{A character string or a vector specifying names of people responsible for identifying the organism. Default value is NULL.}

\item{seq.source}{A character string or a vector specifying data portals from where the (sequence) data was mined. Default value is NULL.}

\item{marker}{A character string or a vector specifying  of gene names. Default value is NULL.}

\item{collection.period}{A Date or a vector of two values specifying the collection period range (start, end). Values separated by comma. Default value is NULL.}

\item{basecount}{A number or a vector of two numbers indicating the base pairs number range. Values separated by a comma. Default value is NULL.}

\item{altitude}{A number or a vector of two numbers indicating the altitude range. Values separated by a comma. Default value is NULL.}

\item{depth}{A number or a vector of two numbers indicating the depth range. Values separated by a comma. Default value is NULL.}
}
\description{
Filters for specific parameters to customize the search for private data
}
\keyword{internal}
