% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.connectr.public.R
\name{bold.connectr.public}
\alias{bold.connectr.public}
\title{Retrieve publicly available data from the BOLD database}
\usage{
bold.connectr.public(
  taxonomy = NULL,
  geography = NULL,
  bins = NULL,
  ids = NULL,
  datasets = NULL,
  latitude = NULL,
  longitude = NULL,
  shapefile = NULL,
  institutes = NULL,
  identified.by = NULL,
  seq.source = NULL,
  marker = NULL,
  collection.period = NULL,
  basecount = NULL,
  altitude = NULL,
  depth = NULL,
  fields = NULL,
  export = NULL,
  file.type,
  file.path,
  file.name
)
}
\arguments{
\item{taxonomy}{A single or multiple character vector specifying the taxonomic names at any hierarchical level. Default value is NULL.}

\item{geography}{A single or multiple character vector specifying any of the country/province/state/region/sector/site names/codes. Default value is NULL.}

\item{bins}{A single or multiple character vector specifying the BIN ids. Default value is NULL.}

\item{ids}{A single or multiple character vector specifying either processids or sampleids. Default value is NULL.}

\item{datasets}{A single or multiple character vector specifying dataset codes. Default value is NULL.}

\item{latitude}{A single or a vector of two numbers specifying the latitudinal range. Values should be separated by a comma. Default value is NULL.}

\item{longitude}{A single or a vector of two numbers specifying the longitudinal range. Values should be separated by a comma. Default value is NULL.}

\item{shapefile}{A file path pointing to a shapefile or name of the shapefile imported in the R session. Default value is NULL.}

\item{institutes}{A single or multiple character vector specifying names of institutes. Default value is NULL.}

\item{identified.by}{A single or multiple character vector specifying names of people responsible for identifying the organism. Default value is NULL.}

\item{seq.source}{A single or multiple character vector specifying the data portals from where the (sequence) data was mined. Default value is NULL.}

\item{marker}{A single or multiple character vector specifying  of gene names. Default value is NULL.}

\item{collection.period}{A single or a vector of two date values specifying the collection period range (start, end). Values should be separated by a comma. Default value is NULL.}

\item{basecount}{A single or a vector of two numbers specifying range of basepairs number. Values should be separated by a comma. Default value is NULL.}

\item{altitude}{A single or a vector of two numbers specifying the altitude range. Values should be separated by a comma. Default value is NULL.}

\item{depth}{A single or a vector of two numbers specifying the depth range. Values should be separated by a comma. Default value is NULL.}

\item{fields}{A single or multiple character vector specifying columns needed in the final dataframe. Default value is NULL.}

\item{export}{A logical value specifying whether the output should be exported locally. Default value is FALSE.}

\item{file.type}{A character value specifying the type of file to be exported. Currently ‘.csv’ and ‘.tsv’ options are available}

\item{file.path}{A character value specifying the folder path where the file should be saved.}

\item{file.name}{A character value specifying the name of the exported file.}
}
\value{
A data frame containing all the information related to the query search
}
\description{
Retrieves publicly available data based on taxonomy,geography or ids (processid,sampleid, dataset codes & bin_uri) input.
}
\details{
\code{bold.connectr.public} downloads publicly available data on BOLD. Data can be retrieved by providing either one or a combination of taxonomy, geography, bins, ids or datasets codes. There is no limit on the data that can be downloaded but complex combinations of the search parameters can lead to web-url character length exceeding the predetermined limit (2048 characters). Single parameter searches are exempt from this limit. Multi-parameter searches (taxonomy + geography + bins/ids/datasets; Please see the example: Taxonomy + Geography + BIN id) to obtain a filtered result should be logical otherwise output obtained might either be empty or not correct. Downloaded data can then be filtered on either one or a combination of arguments (such as institutes, identifiers, altitude, depth etc.). Using the fields argument allows users to select specific columns for inclusion in the final data frame, though, processids and sampleids, are included by default. If the \code{fields} argument is left as NULL all columns will be downloaded. The names of the columns in the downloaded data are those specified in the the \code{bold.fields.info}. Note that some values or fields might currently be unavailable but may be accessible future.
}
\examples{

\dontrun{
# Taxonomy
bold.data<-bold.connectr.public(taxonomy = "Panthera leo")
head(bold.data,10)

# Taxonomy and Geography
bold.data.taxo.geo<-bold.connectr.public(taxonomy = "Panthera uncia",geography = "India")
head(bold.data.taxo_geo,10)

# Taxonomy, Geography and BINs
bold.data.taxo.geo.bin<-bold.connectr.public("Panthera leo", "India",bins=c("BOLD:AAD6819"))
}

}
