% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bold.fetch.R
\name{bold.fetch}
\alias{bold.fetch}
\title{Retrieve all data from the BOLD database}
\usage{
bold.fetch(
  get_by,
  identifiers,
  cols = NULL,
  export = NULL,
  na.rm = FALSE,
  filt_taxonomy = NULL,
  filt_geography = NULL,
  filt_latitude = NULL,
  filt_longitude = NULL,
  filt_shapefile = NULL,
  filt_institutes = NULL,
  filt_identified.by = NULL,
  filt_seq_source = NULL,
  filt_marker = NULL,
  filt_collection_period = NULL,
  filt_basecount = NULL,
  filt_altitude = NULL,
  filt_depth = NULL
)
}
\arguments{
\item{get_by}{The parameter on which the data should be fetched (“processid”, “sampleid”, "bin_uri", "dataset_codes" or "project_codes")}

\item{identifiers}{A vector (or a data frame column) pointing to the \code{get_by} parameter specified.}

\item{cols}{A single or multiple character vector specifying columns needed in the final dataframe. Default value is NULL.}

\item{export}{A logical value specifying whether file path where the file should be exported locally along with the name of the file. Default value is NULL.}

\item{na.rm}{A logical value specifying whether NA values should be removed from the BCDM dataframe. Default value is FALSE.}

\item{filt_taxonomy}{A single or multiple character vector of taxonomic names at any hierarchical level. Default value is NULL.}

\item{filt_geography}{A single or multiple character vector specifying any of the country/province/state/region/sector/site names/codes. Default value is NULL.}

\item{filt_latitude}{A single or a vector of two numbers specifying the latitudinal range in decimal degrees. Values should be separated by a comma. Default value is NULL.}

\item{filt_longitude}{A single or a vector of two numbers specifying the longitudinal range in decimal degrees. Values should be separated by a comma. Default value is NULL.}

\item{filt_shapefile}{A file path pointing to a shapefile or name of the shapefile (.shp) imported in the R session. Default value is NULL.}

\item{filt_institutes}{A single or multiple character vector specifying names of institutes. Default value is NULL.}

\item{filt_identified.by}{A single or multiple character vector specifying names of people responsible for identifying the organism. Default value is NULL.}

\item{filt_seq_source}{A single or multiple character vector specifying the data portals from where the (sequence) data was mined. Default value is NULL.}

\item{filt_marker}{A single or multiple character vector specifying of gene names. Default value is NULL.}

\item{filt_collection_period}{A single or a vector of two date values specifying the collection period range (start, end). Values should be separated by a comma. Default value is NULL.}

\item{filt_basecount}{A single or a vector of two numbers specifying range of basepairs number. Val- ues should be separated by a comma. Default value is NULL.}

\item{filt_altitude}{A single or a vector of two numbers specifying the altitude range in meters. Values should be separated by a comma. Default value is NULL.}

\item{filt_depth}{A single or a vector of two numbers specifying the depth range. Values should be separated by a comma. Default value is NULL.}
}
\value{
A data frame containing all the information related to the processids/sampleids and the filters applied (if/any).
}
\description{
Retrieves public and private user data based on different parameters (processid, sampleid, dataset or project codes & bin_uri) input.
}
\details{
\code{bold.fetch} retrieves both public as well as private user data, where private data refers to data that the user has permission to access. The data is downloaded in the Barcode Core Data Model (BCDM) format. It supports effective download data in bulk using search parameters \code{query.params} such as ‘processids’, ‘sampleids’, ‘bin_uri’ and ‘dataset codes’. Data input can be either through a path to a flat file with extensions like \code{.csv/.tsv/.txt} or a R \code{data.frame} object. The import process assumes that the input data includes a header. Users must specify only one of the \code{query.params} at a time for retrieval. Multi-parameter searches combining fields like ‘processids’+ ‘sampleids’ + ‘bin_uri’ are not supported, regardless of the parameters available. The \code{filt_} or filter parameter arguments provide further data sorting by which a specific user defined data can be obtained. Note that any/all \code{filt_}argument names must be written explicitly to avoid any errors (Ex. \code{filt_institutes} = ’CBG’ instead of just ’CBG’). Using the \code{filt_fields} argument allows users to select specific columns for inclusion in the final data frame, though, processids and sampleids, are included by default. If this argument is left as NULL all columns will be downloaded. There is no upper limit to the volume of data that can be retrieved, however, this depends on the user’s internet connection and computer specifications. The \code{api_key} is a UUID v4 hexadecimal string obtained upon request from BOLD at support@boldsystems.org and is valid for one year, requiring renewal thereafter. The names of the columns in the downloaded data correspond to those specified in bold.fields.info. It is important to correctly match the \code{query.param} and \code{param.index} to avoid getting any errors. Note that some values or fields might currently be unavailable but may be accessible future.
}
\examples{
\dontrun{
Test data with processids
data(test.data)

# apikey is loaded in the R environment using `bold.apikey` function
# Key(token) should be pasted in the function between single quotes.
bold.apikey()

# With processids
res <- bold.fetch(get_by = "processid",
                  identifiers = test.data$processid)


# With sampleids
res<-bold.fetch(get_by = "sampleid",
                identifiers = test.data$sampleid)

# With datasets (publicly available dataset provided)
res<-bold.fetch(get_by = "dataset_codes",
                identifiers = "DS-IBOLR24")

## Using filters

# Geography
res <- bold.fetch(get_by = "processid",
                  identifiers = test.data$processid,
                  filt_geography = "Churchill")

# Sequence length
res <- bold.fetch(get_by = "processid",
                  identifiers = test.data$processid,
                  filt_basecount = c(500,600))

# Gene marker & sequence length
res<-bold.fetch(get_by = "processid",
                identifiers = test.data$processid,
                filt_marker = "COI-5P",
                filt_basecount = c(500, 600))
}

}
