% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/helper.connect.pvt.4_allparams.transform.cols.w.multirec.R
\name{bold.multirecords.set.csv}
\alias{bold.multirecords.set.csv}
\title{Helper function bold.multirecords.set.csv}
\usage{
bold.multirecords.set.csv(edited.json)
}
\arguments{
\item{edited.json}{edited json file obtained after the fetch functions for any of the id types}
}
\description{
Helper function bold.multirecords.set.csv
}
\keyword{internal}
