% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.column_presets.R
\docType{data}
\name{bold_fields}
\alias{bold_fields}
\title{Define and select column presets from the BCDM data}
\format{
An object of class \code{data.frame} with 76 rows and 3 columns.
}
\usage{
bold_fields
}
\description{
Define and select column presets from the BCDM data
}
\keyword{internal}
