% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.connect.pvt.fetch.bins.R
\name{fetch.bold.bins}
\alias{fetch.bold.bins}
\title{fetch processid data using bin ids}
\usage{
fetch.bold.bins(data.input, param.index, api_key)
}
\arguments{
\item{data.input}{input data}

\item{param.index}{the column number in the data which has bin_uri data}

\item{api_key}{API key required to fetch the data}
}
\description{
fetch processid data using bin ids
}
