% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.connect.pvt.fetch.datasets.R
\name{fetch.bold.datasets}
\alias{fetch.bold.datasets}
\title{fetch processids data using the dataset codes}
\usage{
fetch.bold.datasets(data.input, param.index, api_key)
}
\arguments{
\item{data.input}{input data}

\item{param.index}{the column number in the data which has the dataset_codes}

\item{api_key}{API key required to fetch the data}
}
\description{
fetch processids data using the dataset codes
}
\keyword{internal}
