% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.fetch.bold.data.allparams.R
\name{fetch.bold.processid}
\alias{fetch.bold.processid}
\title{fetch data as per the query parameters (processids,sampleids,bin_uri,datasets,projects)}
\usage{
fetch.bold.processid(data.input, param.index, api_key)
}
\description{
fetch data as per the query parameters (processids,sampleids,bin_uri,datasets,projects)
}
\keyword{internal}
