% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.connect.pvt.fetch.sampleids.R
\name{fetch.bold.sampleid}
\alias{fetch.bold.sampleid}
\title{bold.fetch.data sampleid}
\usage{
fetch.bold.sampleid(data.input, param.index, api_key)
}
\arguments{
\item{data.input}{input data}

\item{param.index}{the column number in the data which has bin_uri data}

\item{api_key}{API key required to fetch the data}
}
\description{
bold.fetch.data sampleid
}
\keyword{internal}
