% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.connect.pvt.4_allparams.json2df.R
\name{fetch.data}
\alias{fetch.data}
\title{Helper Function: Fetching the data based on the temp file}
\usage{
fetch.data(result)
}
\arguments{
\item{result}{the output of any of the id fetch functions}
}
\description{
Helper Function: Fetching the data based on the temp file
}
\keyword{internal}
