% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.public.search.public_data.R
\name{fetch.public.data}
\alias{fetch.public.data}
\title{Helper function to retrieve public data}
\usage{
fetch.public.data(query)
}
\arguments{
\item{query}{the output of any of the id fetch functions}
}
\description{
Helper function to retrieve public data
}
\keyword{internal}
