% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.connect.pvt_public.generate_batches_id.R
\name{generate.batches}
\alias{generate.batches}
\title{Helper function to generate batches}
\usage{
generate.batches(data, param.index, batch.size)
}
\arguments{
\item{data}{input data}

\item{param.index}{the column number in the data which has bin_uri data}

\item{batch.size}{Number specifying the number of batches to be created}
}
\description{
Helper function to generate batches
}
\keyword{internal}
