% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.get_data4connect.pvt.fetch.bins.R
\name{get_data_bins}
\alias{get_data_bins}
\title{Helper function using GET to get the processids from bin ids}
\usage{
get_data_bins(data_for_bins, api.key)
}
\arguments{
\item{data_for_bins}{input data containing BIN information}

\item{api.key}{API key required to fetch the data}
}
\description{
Helper function using GET to get the processids from bin ids
}
\keyword{internal}
