% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.get_data4connect.pvt.fetch.dataset.R
\name{get_data_datasets}
\alias{get_data_datasets}
\title{Helper function using GET to get the processids from dataset codes}
\usage{
get_data_datasets(data_for_datasets, api.key)
}
\arguments{
\item{data_for_datasets}{input data containing dataset codes information}

\item{api.key}{API key required to fetch the data}
}
\description{
Helper function using GET to get the processids from dataset codes
}
