% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.fetch.data.allparams.using.GET_API.R
\name{get_data_processid}
\alias{get_data_processid}
\title{GET_data helper functions for ids}
\usage{
get_data_processid(input_data, param.index, api.key)
}
\description{
GET_data helper functions for ids
}
\keyword{internal}
