% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.get_data4connect.pvt.fetch.project.R
\name{get_data_projects}
\alias{get_data_projects}
\title{Helper function using GET to get the processids from project codes}
\usage{
get_data_projects(data_for_projects, api.key)
}
\arguments{
\item{data_for_projects}{input data containing dataset codes information}

\item{api.key}{API key required to fetch the data}
}
\description{
Helper function using GET to get the processids from project codes
}
\keyword{internal}
