% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.fetch.functions.4.datawrangling.R
\name{id.files}
\alias{id.files}
\title{Helper Functions for fetching data}
\usage{
id.files(ids)
}
\description{
Helper Functions for fetching data
}
\keyword{internal}
