% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.column_presets.R
\name{presets}
\alias{presets}
\title{Define and select column presets from the BCDM data}
\usage{
presets(col_groups)
}
\description{
Define and select column presets from the BCDM data
}
\keyword{internal}
