% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/helper.connect.pvt.4_allparams.reassign.data.type.R
\name{reassign.data.type}
\alias{reassign.data.type}
\title{Helper function reassign.data.type}
\usage{
reassign.data.type(x)
}
\arguments{
\item{x}{Data frame}
}
\description{
Helper function reassign.data.type
}
\keyword{internal}
