% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.4.bold.analyze.diversity.R
\name{richness_profile}
\alias{richness_profile}
\title{Helper functions for generating a diversity profile}
\usage{
richness_profile(df)
}
\description{
Helper functions for generating a diversity profile
}
\keyword{internal}
