% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.4.bold.data.summarize.R
\name{taxon_hierarchy_count}
\alias{taxon_hierarchy_count}
\title{Helper functions for creating different data summaries}
\usage{
taxon_hierarchy_count(bold_df)
}
\description{
Helper functions for creating different data summaries
}
\keyword{internal}
