% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.data2.R
\docType{data}
\name{test.data2}
\alias{test.data2}
\title{Example Barcode Core Data Model (BCDM) dataset}
\format{
A data frame with 78 rows and 72 columns.

The returned object follows the Barcode Core Data Model (BCDM) standard and
contains specimen, taxonomic, geographic, collection, and sequence metadata.
Fields correspond to the official BCDM field definitions (\url{https://github.com/boldsystems-central/BCDM})
}
\source{
\url{https://doi.org/10.5883/DP-BOLD_Public.30-Jun-2026}
}
\usage{
test.data2
}
\description{
The \code{test.data2} object is an example of a BCDM output obtained using the
\code{bold.fetch} function. It contains public BOLD records for the freshwater
crustacean genus \emph{Streptocephalus}, consisting of 78 records. The dataset
was obtained from the BOLD Public Data Package released on June 30, 2026.
}
\keyword{datasets}
