% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.seq.R
\name{align.seq}
\alias{align.seq}
\title{Transform and align the sequence data retrieved from BOLD}
\usage{
align.seq(
  bold.df,
  marker = NULL,
  name.fields = NULL,
  file.path = NULL,
  file.name = NULL,
  raw.fas = FALSE
)
}
\arguments{
\item{bold.df}{A data frame obtained from \code{\link[=bold.connectr]{bold.connectr()}} or \code{\link[=bold.connectr.public]{bold.connectr.public()}}.}

\item{marker}{A single or multiple character vector specifying the gene marker for which the output is generated. Default is NULL (all data is used).}

\item{name.fields}{A single or multiple character vector specifying the column headers which to be should be used to name each sequence in the fasta file. Default is NULL in which case, only the processid is used as a name.}

\item{file.path}{A character value specifying the folder path where the file should be saved. Default value is NULL.}

\item{file.name}{A character value specifying the name of the exported file. Default value is NULL.}

\item{raw.fas}{A logical input to specify whether an unaligned(raw) ‘fasta’ file should be created. Default value is FALSE.}
}
\value{
An 'output' list containing:
\itemize{
\item msa.result = A \code{DNAStringSet} object of the multiple sequence alignment.
\item ape_obj = A \code{DNAbin} object of the unaligned sequences.
\item seq.df = A two-column data frame with unaligned sequences in one column and its name in the other.
\item raw.fas = TRUE: a \code{.fas} file of unaligned sequences.
}
}
\description{
Function designed to transform and align the data retrieved from the functions \code{bold.connectr} and \code{bold.connectr.public}
}
\details{
\code{align.seq} retrieves the sequence information obtained using \code{bold.connectr} or \code{bold.connectr.public} functions and performs a multiple sequence alignment using ClustalOmega. multiple sequence alignment on it. It utilizes the msa::msa() function with default settings, specifically calling msa::msaClustalOmega().File path and file name need to be provided for if raw.fas=TRUE. marker name provided must match with the standard marker names available in BOLD. Name for individual sequences in the output can be customized by using the name.fields argument. If more than one field is specified, the name will follow the sequence of the fields given in the vector.Performing a multiple sequence alignment on large sequence data might slow the system. Additionally,users are responsible for   verifying the sequence quality and integrity, as the function does not provide any checks on issues like STOP codons and indels within the data.
\emph{Note: } This function is currently an internal function with documentation. sers are required to install and load the \code{Biostrings} and \code{msa} packages before running this function that are maintained by \code{Bioconductor}.
}
\examples{
\dontrun{
# Download the data
seq<-bold.connectr.public(taxonomy = c("Oreochromis tanganicae","Oreochromis karongae"))

# Align the data (using species", bin_uri & country.ocean as a composite name for each sequence)
seq.align<-align.seq(seq,name.fields = c("species","bin_uri"),marker="COI-5P")

# Dataframe of the sequences (not aligned) with their corresponding names
head(seq.align$seq.df)

#ape DNAbin object
seq.align$ape_obj

#A DNAStringSet object of the Multiple sequence alignment
seq.align$msa.result
 }

# Functions used by `align.seq` (for which the users need to install `Biostrings` & `msa`).
#importFrom Biostrings DNAStringSet
#importFrom msa msa
#importFrom msa msaClustalOmega
#importFrom ape read.dna
#importFrom ape write.FASTA
#importFrom methods as


}
