% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.alphadiv.R
\name{analyze.alphadiv}
\alias{analyze.alphadiv}
\title{Create an alpha diversity profile of the retrieved data}
\usage{
analyze.alphadiv(
  bin.comm,
  plot.curve = FALSE,
  curve.index = NULL,
  curve.xval = NULL,
  preston.res = FALSE,
  pres.plot.y.label = NULL
)
}
\arguments{
\item{bin.comm}{the site X species like output from the \code{gen.comm.mat} function.}

\item{plot.curve}{A logical value specifying whether an accumulation curve should be plotted.Default value is FALSE.}

\item{curve.index}{A character value specifying which index should be used for the \code{curve.index} argument. Default value is NULL.}

\item{curve.xval}{A character value specifying whether samples (Samples) or individuals (Individuals) should be used against the \code{curve.index}.Default value is NULL.}

\item{preston.res}{A logical value specifying whether the Preston results should be generated. Default value is FALSE.}

\item{pres.plot.y.label}{A character value specifying the taxonomic category (\code{taxon.rank} in \code{\link[=gen.comm.mat]{gen.comm.mat()}}) which was used to generate the matrix.}
}
\value{
An 'output' list containing containing:
\itemize{
\item richness = A richness profile matrix
\item Shannon_div = Shannon diversity values for the given sites/grids(from \code{gen.comm.mat})
\item richness_plot = A ggplot2 visualization of the richness curve
\item preston.res = a Preston plot numerical data output
\item preston.plot = a ggplot2 visualization of the preston.plot
}
}
\description{
This function analyzes the output from the \code{gen.comm.mat} to provide a richness and alpha diversity profile of the downloaded data.
}
\details{
\code{analyze.alphadiv} estimates the richness and calculates the shannon diversity indexes from BIN counts or presence-absence data at the user-specified taxonomic level using the \code{gen.comm.mat()} output. It generates richness profiles using \code{\link[BAT:alpha.accum]{BAT::alpha.accum()}} and Preston and Shannon diversity analyses using \code{\link[vegan:fisherfit]{vegan::prestondistr()}} and \code{\link[vegan:diversity]{vegan::diversity()}} respectively and depend on the input data (true abundances vs counts vs incidences). Preston plots are created using the data from the \code{prestondistr} results in \code{ggplot2} featuring cyan bars for observed species (or equivalent taxonomic group) and orange dots for expected counts.
\verb{\\emph\{Note:\}} Results, including species counts, adapt based on taxonomic rank used in \code{gen.comm.mat()} although the output label remains ‘species’ in some instances (\code{preston.res}).
}
\examples{
\dontrun{
# Download data from BOLD (removing species with blanks)
comm.mat.data<-bold.connectr.public(taxonomy = "Poecilia")

# Remove rows which have no species data
comm.mat.data<-comm.mat.data[!comm.mat.data$species=="",]

# Generate the community matrix based on grids
comm.data.grid<-gen.comm.mat(comm.mat.data,taxon.rank="species",site.cat='country.ocean')

grid.data<-comm.data.grid$comm.matrix

# Diversity results with estimation curve and without preston results
div.res1<-analyze.alphadiv(grid.data,plot.curve=TRUE,curve.index="Jack1ab",curve.xval = "Sampl")

# Richness estimations
head(div.res1$richness)

# Richness plot
div.res1$richness_plot

# Shannon diversity
head(div.res1$Shannon_div)


# Diversity results without estimation curve and with preston results
div.res2<-analyze.alphadiv(grid.data,preston.res = TRUE,pres.plot.y.label = "species")

# Preston results
div.res2$preston.res

# Preston plot
div.res2$preston.plot
}

}
