% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.betadiv.R
\name{analyze.betadiv}
\alias{analyze.betadiv}
\title{Create a beta diversity profile of the retrieved data}
\usage{
analyze.betadiv(
  bin.comm,
  index,
  pre.abs = FALSE,
  heatmap = FALSE,
  component,
  grids = FALSE,
  grids.df = NULL
)
}
\arguments{
\item{bin.comm}{the site X species like output from the \code{gen.comm.mat} function.Default is FALSE}

\item{index}{A character vector specifying the type of beta diversity index ('jaccard' or 'sorenson' currently available)}

\item{pre.abs}{A logical value specifying if input data is presence-absence. Default is FALSE.}

\item{heatmap}{A logical value specifying whether a heatmap of the beta diversity values should be plotted.Default is FALSE.}

\item{grids}{A logical value specifying Whether the community matrix is generated using grids. Default is FALSE.}

\item{grids.df}{If grids = TRUE, a 'sf' grid data frame generated along with the community matrix using the\code{gen.comm.mat} function.Default is NULL}
}
\value{
A data frame containing all the information related to the processids/sampleids and the filters applied (if/any)
output$total.beta = beta.total
output$replace = beta.replace
output$richnessd = beta.richnessd
output$heatmap.viz = heatmap_final
}
\description{
Function analyzes the output from the \code{gen.comm.mat} to provide a beta diversity (and its partitions) profile of the downloaded data
}
\details{
analyze.betadiv calculates a sorenson or jaccard beta dissimilarity matrix using the \code{gen.comm.mat} output.
}
