% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.betadiv.R
\name{analyze.betadiv}
\alias{analyze.betadiv}
\title{Create a beta diversity profile of the retrieved data}
\usage{
analyze.betadiv(
  bin.comm,
  index,
  pre.abs = FALSE,
  heatmap = FALSE,
  component,
  grids = FALSE,
  grids.df = NULL
)
}
\arguments{
\item{bin.comm}{the site X species like output from the \code{gen.comm.mat} function.Default value is FALSE.}

\item{index}{A character vector specifying the type of beta diversity index ('jaccard' or 'sorenson' available).}

\item{pre.abs}{A logical value specifying whether the input data is presence-absence. Default value is FALSE.}

\item{heatmap}{A logical value specifying whether a heatmap of the beta diversity values should be plotted.Default value is FALSE.}

\item{component}{A character value specifying which beta diversity component should be used for the heatmap. Default value is NULL.}

\item{grids}{A logical value specifying Whether the community matrix is generated using grids. Default value is FALSE.}

\item{grids.df}{If \code{grids} = TRUE, a \code{sf} grid data frame generated along with the community matrix using the \code{\link[=gen.comm.mat]{gen.comm.mat()}} function.Default value is NULL.}
}
\value{
An 'output' list containing:
\itemize{
\item total.beta = beta.total
\item replace = beta.replace (replacement)
\item richnessd = beta.richnessd (richness difference)
\item heatmap.viz = heatmap_final
}
}
\description{
This function analyzes the output from the \code{gen.comm.mat} to provide a beta diversity (and its partitions) profile of the downloaded data.
}
\details{
\code{analyze.betadiv} calculates either a Sørensen or Jaccard beta dissimilarity indexes  using the \code{gen.comm.mat()} output. It also generates matrices of ’species replacement’ and ’richness difference’ components of the total beta diversity. These values are calculated using \code{\link[BAT:beta]{BAT::beta()}} function, which partitions the data using the Podani & Schmera (2011)/Carvalho et al. (2012) approach. Additionally, a corresponding \code{heatmap} can be obtained when \code{heatmap}=TRUE. For grid-based heatmaps, grids are arranged based on their centroid distances, placing nearest grids closest together. For site categories, heatmap labels are arranged alphabetically. Note that grid- based heatmaps require \code{grids} = TRUE and a spatial dataframe  sf-object \code{grid.df} generated from the \code{gen.comm.mat} function, to be provided.
}
\examples{
\dontrun{
#Download data from BOLD (removing species with blanks)
comm.mat.data<-bold.connectr.public(taxonomy = "Poecilia")

#Generate the community matrix based on grids
comm.data.beta<-gen.comm.mat(comm.mat.data,taxon.rank="species",site.cat = "country.ocean")

beta.data<-comm.data.beta$comm.matrix

#beta diversity without the heatmaps
beta.div.res<-analyze.betadiv(beta.data,index="sorenson")

#Total diversity
beta.div.res$total.beta

#Replacement
beta.div.res$replace

#Richness difference
beta.div.res$richnessd


#beta diversity with the heatmaps
beta.div.res2<-analyze.betadiv(beta.data,index="sorenson",heatmap = TRUE,component = "total")

#Total diversity
beta.div.res2$total.beta

#Replacement
beta.div.res2$replace

#Richness difference
beta.div.res2$richnessd

#Visualize the heatmap
beta.div.res$heatmap.viz
}

}
\references{
Carvalho, J.C., Cardoso, P. & Gomes, P. (2012) Determining the relative roles of species replacement and species richness differences in generating beta-diversity patterns. Global Ecology and Biogeography, 21, 760-771.

Podani, J. & Schmera, D. (2011) A new conceptual and methodological framework for exploring and explaining pattern in presence-absence data. Oikos, 120, 1625-1638.#'
}
