% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.betadiv.R
\name{analyze.betadiv}
\alias{analyze.betadiv}
\title{Create a beta diversity profile of the retrieved data}
\usage{
analyze.betadiv(
  bin.comm,
  index,
  pre.abs = FALSE,
  heatmap = FALSE,
  component,
  grids = FALSE,
  grids.df = NULL
)
}
\arguments{
\item{bin.comm}{the site X species like output from the \code{gen.comm.mat} function.Default value is FALSE.}

\item{index}{A character vector specifying the type of beta diversity index ('jaccard' or 'sorenson' currently available).}

\item{pre.abs}{A logical value specifying whether the input data is presence-absence. Default value is FALSE.}

\item{heatmap}{A logical value specifying whether a heatmap of the beta diversity values should be plotted.Default value is FALSE.}

\item{component}{A character value specifying which beta diversity component should be used for the heatmap. Default value is NULL.}

\item{grids}{A logical value specifying Whether the community matrix is generated using grids. Default value is FALSE.}

\item{grids.df}{If \code{grids} = TRUE, a \link{sf} grid data frame generated along with the community matrix using the \code{\link[=gen.comm.mat]{gen.comm.mat()}} function.Default value is NULL.}
}
\value{
A list containing:
\itemize{
\item output$total.beta = beta.total
\item output$replace = beta.replace (replacement)
\item output$richnessd = beta.richnessd (richness difference)
\item output$heatmap.viz = heatmap_final
}
}
\description{
This function analyzes the output from the \code{gen.comm.mat} to provide a beta diversity (and its partitions) profile of the downloaded data.
}
\details{
\code{analyze.betadiv} calculates either a sorenson or jaccard beta dissimilarity using the \code{\link[=gen.comm.mat]{gen.comm.mat()}} output. It also generates matrices of 'species replacement' and 'richness difference' components of the total beta diversity. The values are calculated using \code{\link[BAT:beta]{BAT::beta()}} function which partitions the data using the Podani approach. A corresponding 'heatmap' can also be obtained when \code{heatmap}=TRUE. In case of grid based heatmaps, grids are arranged on the heatmap based on their centroid distances (i.e. nearest grids are placed closest). For site categories, the heatmap labels are arranged alphabetically. Grid based heatmaps can only be generated when \code{grids} = TRUE and a \link{sf} 'grid.df' which is generated from the \code{gen.comm.mat} function is provided to the function.
}
\examples{

#Download data from BOLD (removing species with blanks)
comm.mat.data<-bold.connectr.public(taxonomy = "Poecilia")

#Generate the community matrix based on grids
comm.data.beta<-gen.comm.mat(comm.mat.data,taxon.rank="species",site.cat = "country.ocean")

beta.data<-comm.data.beta$comm.matrix

#beta diversity without the heatmaps
beta.div.res<-analyze.betadiv(beta.data,index="sorenson")

#Total diversity
beta.div.res$total.beta

#Replacement
beta.div.res$replace

#Richness difference
beta.div.res$richnessd


#beta diversity with the heatmaps
beta.div.res2<-analyze.betadiv(beta.data,index="sorenson",heatmap = TRUE,component = "total")

#Total diversity
beta.div.res2$total.beta

#Replacement
beta.div.res2$replace

#Richness difference
beta.div.res2$richnessd

#Visualize the heatmap
beta.div.res$heatmap.viz

}
