% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.seq.R
\name{analyze.seq}
\alias{analyze.seq}
\title{Analyze and visualize the output from the \code{align.seq}}
\usage{
analyze.seq(
  aligned.seq,
  dist.model,
  clus = c("nj", "njs"),
  tree.export = FALSE,
  file.path = NULL,
  file.name = NULL,
  plot = FALSE,
  ...
)
}
\arguments{
\item{aligned.seq}{A DNAStringset multiple sequence alignment object returned by \code{align.seq} function}

\item{dist.model}{A character string specifying the model to generate the distances}

\item{clus}{A character vector specifying either"nj" (neighbour joining) or "njs" (neighbour joining with NAs) clustering algorithm}

\item{tree.export}{Logical value specifying whether newick tree should be generated and exported. Default value is FALSE,}

\item{file.path}{A character value specifying the folder path where the file should be saved}

\item{file.name}{A character value specifying the name of the exported file.}

\item{plot}{Logical value specifying if a neighbour joining plot should be generated. Default value is FALSE}

\item{...}{additional arguments that can be provided to the \code{dist.dna} function from ape}
}
\value{
dist_mat = A distance matrix based on the model selected
base_freq = Overall base frequencies of the 'align.seq' result
newick_tree = NJ/NJS tree in a newick format (only if tree.export=TRUE)
plot = Neighbour Joining clustering visualization (if plot=TRUE)
}
\description{
This function provides a distance matrix and a NJ clustering of the multiple sequence alignment output of the \code{align.seq}.
}
\details{
‘analyze.seq’ analyzes the output of the \code{align.seq} function to generate a distance matrix using the models available in the \code{dist.dna} function of the ape library. The function also has an option to plot the distances as a basic Neighbour joining cluster. Additionally, the function provides base frequencies and a option to export the trees in 'newick' format.
}
