% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.4.bold.public.search.R
\docType{data}
\name{base_url_parse}
\alias{base_url_parse}
\title{Helper functions for bold.public.search}
\format{
An object of class \code{character} of length 1.
}
\usage{
base_url_parse
}
\description{
Helper functions for bold.public.search
}
\keyword{internal}
