% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.analyze.align.R
\name{bold.analyze.align}
\alias{bold.analyze.align}
\title{Transform and align the sequence data retrieved from BOLD (Internal function)}
\usage{
bold.analyze.align(
  bold.df,
  marker = NULL,
  align.method,
  seq.name.fields = NULL,
  ...
)
}
\arguments{
\item{bold.df}{A data frame obtained from \code{\link[=bold.fetch]{bold.fetch()}}.}

\item{marker}{A single or multiple character vector specifying the gene marker for which the output is generated. Default is NULL (all data is used).}

\item{align.method}{Character vector specifying the type of multiple sequence alignment algorithm to be used. Default value is ClustalOmega.}

\item{seq.name.fields}{A single or multiple character vector specifying the column headers which to be should be used to name each sequence in the fasta file. Default is NULL in which case, only the processid is used as a name.}

\item{...}{additional arguments that can be passed to \code{msa::msa()} function.}
}
\value{
An 'output' list containing:
\itemize{
\item bold.df.mod = A modified bold BCDM data frame with two additional columns ('aligned_seq' and 'sequence_name').
}
}
\description{
Function designed to transform and align the data retrieved from the functions \code{bold.fetch}.
}
\details{
\code{bold.analyze.align} retrieves the sequence information obtained using \code{bold.fetch} function and performs a multiple sequence alignment using ClustalOmega algorithm. It utilizes the msa::msa() function with default settings but additional arguments can be passed via the \code{...} argument. Marker name provided must match with the standard marker names available on the BOLD webpage. Name for individual sequences in the output can be customized by using the \code{sequence.name.fields} argument. If more than one field is specified, the name will follow the sequence of the fields given in the vector. Performing a multiple sequence alignment on large sequence data might slow the system. Additionally,users are responsible for verifying the sequence quality and integrity, as the function does not provide any checks on issues like STOP codons and indels within the data. The output of this function is a modified BCDM dataframe having two extra columns, one of the aligned sequence and the other of the name given to that sequence.

\emph{Note: } This function is currently an internal function with documentation and can be accessed only by using the \code{:::} operator  (\code{BOLDconnectR:::bold.analyze.align}). Users are required to install and load the \code{Biostrings} and \code{msa} packages before running this function that are maintained by \code{Bioconductor}.
}
\examples{
\dontrun{

# Search for ids
seq.data.ids<-bold.public.search(taxonomy = c("Oreochromis tanganicae","Oreochromis karongae"))

# Fetch the data using the ids
seq.data<-bold.fetch(param.data = seq.data.ids,query.param = "processid",param.index = 1,apikey)

# msa is reuqired for this function to run. msa is installed using BiocManager
if (!requireNamespace("BiocManager", quietly=TRUE))
install.packages("BiocManager")
BiocManager::install("msa")
BiocManager::install("Biostrings")
library(msa)
library(Biostrings)

# Align the data (using species", bin_uri & country.ocean as a composite name for each sequence)
seq.align<-BOLDconnectR:::bold.analyze.align(seq.data, seq.name.fields = c("bin_uri"))

# Dataframe of the sequences (aligned) with their corresponding names
 head(seq.align)
 }

}
