% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.analyze.diversity.R
\name{bold.analyze.diversity}
\alias{bold.analyze.diversity}
\title{Create an biodiversity profile of the retrieved data}
\usage{
bold.analyze.diversity(
  bold.df,
  taxon.rank,
  taxon.name = NULL,
  site.cat = NULL,
  grids.cat = FALSE,
  gridsize = NULL,
  presence.absence = FALSE,
  richness.res = FALSE,
  rich.plot.curve = FALSE,
  rich.curve.estimatr = NULL,
  rich.curve.x.axis = NULL,
  shannon.res = FALSE,
  preston.res = FALSE,
  pres.plot.y.label = NULL,
  beta.res = FALSE,
  beta.index = NULL
)
}
\arguments{
\item{bold.df}{A data frame obtained from \code{\link[=bold.fetch]{bold.fetch()}}.}

\item{taxon.rank}{A single character value specifying the taxonomic hierarchical rank. Needs to be provided by default.}

\item{taxon.name}{A single or multiple character vector specifying the taxonomic names associated with the ‘taxon.rank’. Default value is NULL.}

\item{site.cat}{A single or multiple character vector specifying the countries for which a community matrix should be created. Default value is NULL.}

\item{grids.cat}{A logical value specifying Whether the community matrix should be based on grids as ‘sites’. Default value is NULL.}

\item{gridsize}{A numeric value of the size of the grid if the grids=TRUE;Size is in sq.m. Default value is NULL.}

\item{presence.absence}{A logical value specifying whether the generated matrix should be converted into a 'presence-absence' matrix.}

\item{richness.res}{A logical value specifying whether richness results should be calculated. Default value is FALSE.}

\item{rich.plot.curve}{A logical value specifying whether an accumulation curve should be plotted.Default value is FALSE.}

\item{rich.curve.estimatr}{A character value specifying which index should be used for the \code{curve.index} argument. Default value is NULL.}

\item{rich.curve.x.axis}{A character value specifying whether samples (Samples) or individuals (Individuals) should be used against the \code{curve.index}.Default value is NULL.}

\item{shannon.res}{A logical value specifying whether the Shannon diversity results should be generated. Default value is FALSE.}

\item{preston.res}{A logical value specifying whether the Preston results should be generated. Default value is FALSE.}

\item{pres.plot.y.label}{A character value specifying the taxonomic category (\code{taxon.rank} in \code{\link[=gen.comm.mat]{gen.comm.mat()}}) which was used to generate the matrix.}

\item{beta.res}{A logical value specifying whether beta diversity results should be calculated. Default value is FALSE.}

\item{beta.index}{A character vector specifying the type of beta diversity index ('jaccard' or 'sorenson' available).}
}
\value{
An 'output' list containing containing:
\itemize{
\item comm.matrix = site X species like matrix required for the biodiversity results
\item richness = A richness profile matrix
\item Shannon_div = Shannon diversity values for the given sites/grids(from \code{gen.comm.mat})
\item richness_plot = A ggplot2 visualization of the richness curve
\item preston.res = a Preston plot numerical data output
\item preston.plot = a ggplot2 visualization of the preston.plot
\item total.beta = beta.total
\item replace = beta.replace (replacement)
\item richnessd = beta.richnessd (richness difference)
}
}
\description{
This function creates a biodiversity profile of the downloaded data using \code{\link[=bold.fetch]{bold.fetch()}}.
}
\details{
\code{bold.analyze.diversity} estimates the richness, shannon diversity and beta diversity  from the BIN counts or presence-absence data. The function internally converts the downloaded BCDM data into a community matrix (~site X species) which is also generated as a part of the output.\code{grids.cat} converts the Coordinate Reference System (CRS) of the data to a ‘Mollweide' projection by which distance based grid can be correctly specified. A cell id is also given to each grid with the lowest number assigned to the lowest latitudinal point in the dataset. It uses this matrix to create richness profiles using \code{\link[BAT:alpha.accum]{BAT::alpha.accum()}} and Preston and Shannon diversity analyses using \code{\link[vegan:fisherfit]{vegan::prestondistr()}} and \code{\link[vegan:diversity]{vegan::diversity()}} respectively. It also generates matrices of ’species replacement’ and ’richness difference’ components of the total beta diversity. These values are calculated using \code{\link[BAT:beta]{BAT::beta()}} function, which partitions the data using the Podani & Schmera (2011)/Carvalho et al. (2012) approach.The results depen on the input data (true abundances vs counts vs incidences) and users should be careful in the subsequent interprettation. Preston plots are created using the data from the \code{prestondistr} results in \code{ggplot2} featuring cyan bars for observed species (or equivalent taxonomic group) and orange dots for expected counts. The argument \code{presence.absence} will convert the counts (or abundances) to 1 and 0. This dataset can then directly be used as the input data for functions from packages like \code{vegan} for biodiversity analyses.
\verb{\\emph\{Note on the community matrix\}}: Instead of species counts (or abundances), values in each cell of this matrix are the counts (or abundances) of a specific BIN from a \code{site.cat} site category or a \code{grids.cat}. These counts can be generated at any taxonomic hierarchical level for a single or multiple taxa (This can also be done for 'bin_uri'; the difference being that the numbers in each cell would be the number of times that respective BIN is found at a particular \code{site.cat} or \code{grids.cat}). \code{site.cat} can be any of the \code{geography} fields (Meta data on fields can be checked using the \code{\link[=bold.fields.info]{bold.fields.info()}}). Alternatively, \code{grids.cat} = TRUE will generate grids based on the BIN occurrence data (latitude, longitude) with the size of the grid determined by the user (in sq.m.) with the \code{gridsize} argument. For \code{grids} generation, rows with no latitude and longitude data are removed (even if a corresponding \code{site.cat} information is available) while NULL entries for \code{site.cat} are allowed if they have a latitude and longitude value (This is done because grids are drawn based on the bounding boxes which only use latitude and longitude values).
\verb{\\emph\{Note:\}} Results, including species counts, adapt based on \code{taxon.rank} argument although the output label remains ‘species’ in some instances (\code{preston.res}).
}
\examples{
\dontrun{
# Download data from BOLD (removing species with blanks)
comm.mat.data<-bold.connectr.public(taxonomy = "Poecilia")

# Remove rows which have no species data
comm.mat.data<-comm.mat.data[!comm.mat.data$species=="",]

# Generate the community matrix based on grids
comm.data.grid<-gen.comm.mat(comm.mat.data,taxon.rank="species",site.cat='country.ocean')

grid.data<-comm.data.grid$comm.matrix

# Diversity results with estimation curve and without preston results
div.res1<-analyze.alphadiv(grid.data,plot.curve=TRUE,curve.index="Jack1ab",curve.xval = "Sampl")

# Richness estimations
head(div.res1$richness)

# Richness plot
div.res1$richness_plot

# Shannon diversity
head(div.res1$Shannon_div)

# Diversity results without estimation curve and with preston results
div.res2<-analyze.alphadiv(grid.data,preston.res = TRUE,pres.plot.y.label = "species")

# Preston results
div.res2$preston.res

# Preston plot
div.res2$preston.plot

 beta.data<-comm.data.beta$comm.matrix

#beta diversity without the heatmaps
beta.div.res<-analyze.betadiv(beta.data,index="sorenson")

#Total diversity
beta.div.res$total.beta

#Replacement
beta.div.res$replace

#Richness difference
beta.div.res$richnessd


}

}
