% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bold.analyze.diversity.R
\name{bold.analyze.diversity}
\alias{bold.analyze.diversity}
\title{Create an biodiversity profile of the retrieved data}
\usage{
bold.analyze.diversity(
  bold_df,
  taxon_rank,
  taxon_name = NULL,
  site_type = c("locations", "grids"),
  location_type = NULL,
  gridsize = NULL,
  presence_absence = FALSE,
  diversity_profile = c("richness", "preston", "shannon", "beta", "all"),
  beta_index = NULL
)
}
\arguments{
\item{bold_df}{A data frame obtained from \code{\link[=bold.fetch]{bold.fetch()}}.}

\item{taxon_rank}{A single character value specifying the taxonomic hierarchical rank. Needs to be provided by default.}

\item{taxon_name}{A single or multiple character vector specifying the taxonomic names associated with the ‘taxon_rank’. Default value is NULL.}

\item{site_type}{A character vector specifying one of two broad categories of \code{sites} considered here (\code{locations} or \code{grids}).}

\item{location_type}{A single character vector specifying the geographic category for which a community matrix should be created. Default value is NULL.}

\item{gridsize}{A numeric value of the size of the grid if location_type=grid; Size is in sq.m. Default value is NULL.}

\item{presence_absence}{A logical value specifying whether the generated matrix should be converted into a ’presence-absence’ matrix.}

\item{diversity_profile}{A character value specifying the type of diversity profile ("richness","preston","shannon","beta")}

\item{beta_index}{A character vector specifying the type of beta diversity index (’jaccard’ or ’sorensen’ available).}
}
\value{
An 'output' list containing containing results based on the profile selected:
#Common to all
\itemize{
\item comm.matrix = site X species like matrix required for the biodiversity results
#1. richness
\item richness = A richness profile matrix
#2. shannon
\item Shannon_div = Shannon diversity values for the given sites/grids (from gen.comm.mat)
#3. preston
\item preston.res = a Preston plot numerical data output
\item preston.plot = a ggplot2 visualization of the preston.plot
#4. beta
\item total.beta = beta.total
\item replace = beta.replace (replacement)
\item richnessd = beta.richnessd (richness difference)
#5. all
\item All of the above results
}
}
\description{
This function creates a biodiversity profile of the downloaded data using \code{\link[=bold.fetch]{bold.fetch()}}.
}
\details{
\code{bold.analyze.diversity} estimates the richness & calculates the Shannon and beta diversity from the BIN counts or presence-absence data. Internally, the function converts the downloaded BCDM data into a community matrix (site X species) which is also generated as a part of the output. \code{taxon_rank} has to be provided by default while \code{taxon_name} if given, will create the matrix for that specific taxon/taxa. \code{site_type} = 'locations' can be used when a profile pertaining to particular geographic category is needed. This category can be specified using the \code{location_type} argument. \code{site_type}= grids creates a grid based on BIN occurrence data (latitude, longitude) with grid size determined by the user in square meters using the \code{gridsize} argument. The the Coordinate Reference System (CRS) of this data is converted to a ‘Mollweide’ projection by which distance-based grid can be correctly specified (Gott III et al. 2007). Each grid is also assigned a cell id, with the lowest number given to the lowest latitudinal point in the dataset. The \code{presence_absence} argument converts the counts (or abundances) to 1s and 0s.
The community matrix is then used to create one of the following \code{diversity_profile} using functions from \code{BAT} and \code{vegan} packages:
\itemize{
\item \code{richness}(\code{BAT::alpha.accum()})
\item \code{preston}(plots and results)(\code{vegan::prestondistr()})
\item \code{shannon}(\code{vegan::diversity()})
\item \code{beta}(\code{BAT::beta()})
The option \code{all} generates results for all of the above.
\code{BAT::alpha.accum()} currently offers various richness estimators, including Observed diversity (Obs); Singletons (S1); Doubletons (S2); Uniques (Q1); Duplicates (Q2); Jackknife1 abundance (Jack1ab); Jackknife1 incidence (Jack1in); Jackknife2 abundance (Jack2ab); Jackknife2 incidence (Jack2in); Chao1 and Chao2. The results depend on the input data (true abundances vs counts vs incidences) and users should be careful in the subsequent interpretation.
Preston plots feature cyan bars for observed species (or equivalent taxonomic group) and orange dots for expected counts.
\code{BAT::beta()} partitions the data using the Podani & Schmera (2011)/Carvalho et al. (2012) approach partitioning the beta diversity into ’species replacement’ and ’richness difference’ components. These results are stored as distance matrices in the output. The type of beta index can be specified using the \code{beta_index} argument.
\emph{Note on the community matrix}: Each cell in this matrix contains the counts (or abundances) of the specimens whose sequences have an assigned BIN, in a given a \code{site_type} (\code{locations} or \code{grids}). These counts can be generated at any taxonomic hierarchical level, applicable to one or multiple taxa including ’bin_uri’. \code{location_type} can refer to any geographic field, and metadata on these fields can be checked using the \code{bold.fields.info()}. Rows lacking latitude and longitude data are removed when \code{site_type} = 'grids', while NULL entries for \code{site_type} = 'locations' are allowed if they have a latitude and longitude value. This is because grids are drawn based on the bounding boxes which only use latitude and longitude values.
\emph{Important Note}: Results, including species counts, adapt based on taxon_rank argument although the output label remains ‘species’ in case of \code{preston} results.
}
}
\examples{
\dontrun{
# Search for ids
comm.mat.data <- bold.public.search(taxonomy = "Poecilia")

# Fetch the data using the ids
#1. api_key must be obtained from BOLD support before usage
#2. the function `bold.apikey` should be used to set the apikey in the global env

BCDMdata <- bold.fetch(get_by = "processid",
                       identifiers = comm.mat.data$processid)

# Remove rows which have no species data
BCDMdata <- BCDMdata[!BCDMdata$species== "",]

#1. Analyze richness data
res.rich <- bold.analyze.diversity(bold_df=BCDMdata,
                                   taxon_rank = "species",
                                   site_type = "locations",
                                   location_type = 'country.ocean',
                                   diversity_profile = "richness")

# Community matrix (BCDM data converted to community matrix)
res.rich$comm.matrix

# richness results
res.rich$richness

#2. Shannon diversity
res.shannon <- bold.analyze.diversity(bold_df=BCDMdata,
                                      taxon_rank = "species",
                                      site_type = "locations",
                                      location_type = 'country.ocean',
                                      diversity_profile = "shannon")

# Shannon diversity results
res.shannon

#3. Preston plots and results
pres.res <- bold.analyze.diversity(bold_df=BCDMdata,
                                   taxon_rank = "species",
                                   site_type = "locations",
                                   location_type = 'country.ocean',
                                   diversity_profile = "preston")

# Preston plot
pres.res$preston.plot

# Preston plot data
pres.res$preston.res

#4. beta diversity
beta.res <- bold.analyze.diversity(bold_df=BCDMdata,
                                   taxon_rank = "species",
                                   site_type = "locations",
                                   location_type = 'country.ocean',
                                   diversity_profile = "beta",
                                   beta_index = "jaccard")

#Total diversity
beta.res$total.beta

#Replacement
beta.res$replace

#Richness difference
beta.res$richnessd
}

}
\references{
Carvalho, J.C., Cardoso, P. & Gomes, P. (2012) Determining the relative roles of species replace- ment and species richness differences in generating beta-diversity patterns. Global Ecology and Biogeography, 21, 760-771.

Podani, J. & Schmera, D. (2011) A new conceptual and methodological framework for exploring and explaining pattern in presence-absence data. Oikos, 120, 1625-1638.

Richard Gott III, J., Mugnolo, C., & Colley, W. N. (2007). Map projections minimizing distance errors. Cartographica: The International Journal for Geographic Information and Geovisualization, 42(3), 219-234.
}
