% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.analyze.diversity.R
\name{bold.analyze.diversity}
\alias{bold.analyze.diversity}
\title{Create an biodiversity profile of the retrieved data}
\usage{
bold.analyze.diversity(
  bold.df,
  taxon.rank,
  taxon.name = NULL,
  site.cat = NULL,
  grids.cat = FALSE,
  gridsize = NULL,
  presence.absence = FALSE,
  richness.res = FALSE,
  rich.plot.curve = FALSE,
  rich.curve.estimatr = NULL,
  rich.curve.x.axis = NULL,
  shannon.res = FALSE,
  preston.res = FALSE,
  pres.plot.y.label = NULL,
  beta.res = FALSE,
  beta.index = NULL
)
}
\arguments{
\item{bold.df}{A data frame obtained from \code{\link[=bold.fetch]{bold.fetch()}}.}

\item{taxon.rank}{A single character value specifying the taxonomic hierarchical rank. Needs to be provided by default.}

\item{taxon.name}{A single or multiple character vector specifying the taxonomic names associated with the ‘taxon.rank’. Default value is NULL.}

\item{site.cat}{A single or multiple character vector specifying the countries for which a community matrix should be created. Default value is NULL.}

\item{grids.cat}{A logical value specifying Whether the community matrix should be based on grids as ‘sites’. Default value is NULL.}

\item{gridsize}{A numeric value of the size of the grid if the grids=TRUE;Size is in sq.m. Default value is NULL.}

\item{presence.absence}{A logical value specifying whether the generated matrix should be converted into a 'presence-absence' matrix.}

\item{richness.res}{A logical value specifying whether richness results should be calculated. Default value is FALSE.}

\item{rich.plot.curve}{A logical value specifying whether an accumulation curve should be plotted.Default value is FALSE.}

\item{rich.curve.estimatr}{A character value specifying which index should be used. Default value is NULL.}

\item{rich.curve.x.axis}{A character value specifying whether samples (Samples) or individuals (Individuals) should be used against the \code{curve.index}.Default value is NULL.}

\item{shannon.res}{A logical value specifying whether the Shannon diversity results should be generated. Default value is FALSE.}

\item{preston.res}{A logical value specifying whether the Preston results should be generated. Default value is FALSE.}

\item{pres.plot.y.label}{A character value specifying the taxonomic category (\code{taxon.rank}) which was used to generate the matrix.}

\item{beta.res}{A logical value specifying whether beta diversity results should be calculated. Default value is FALSE.}

\item{beta.index}{A character vector specifying the type of beta diversity index ('jaccard' or 'sorenson' available).}
}
\value{
An 'output' list containing containing:
\itemize{
\item comm.matrix = site X species like matrix required for the biodiversity results
\item grid_plot = A map of grids based on their centroids
\item richness = A richness profile matrix
\item Shannon_div = Shannon diversity values for the given sites/grids(from \code{gen.comm.mat})
\item richness_plot = A ggplot2 visualization of the richness curve
\item preston.res = a Preston plot numerical data output
\item preston.plot = a ggplot2 visualization of the preston.plot
\item total.beta = beta.total
\item replace = beta.replace (replacement)
\item richnessd = beta.richnessd (richness difference)
}
}
\description{
This function creates a biodiversity profile of the downloaded data using \code{\link[=bold.fetch]{bold.fetch()}}.
}
\details{
\code{bold.analyze.diversity} estimates the richness, shannon diversity and beta diversity  from the BIN counts or presence-absence data. The function internally converts the downloaded BCDM data into a community matrix (~site X species) which is also generated as a part of the output.\code{grids.cat} converts the Coordinate Reference System (CRS) of the data to a ‘Mollweide' projection by which distance based grid can be correctly specified. A cell id is also given to each grid with the lowest number assigned to the lowest latitudinal point in the dataset. It uses this matrix to create richness profiles using \code{\link[BAT:alpha.accum]{BAT::alpha.accum()}} and Preston and Shannon diversity analyses using \code{\link[vegan:fisherfit]{vegan::prestondistr()}} and \code{\link[vegan:diversity]{vegan::diversity()}} respectively. The \code{\link[BAT:alpha.accum]{BAT::alpha.accum()}} currently has many richness estimators (Obs: Observed diversity; S1: Singletons; S2: Doubletons; Q1: Uniques; Q2: Duplicates; Jack1ab: Jackknife1 abundance; Jack1in: Jackknife1 incidence; Jack2ab: Jackknife2 abundance; Jack2in: Jackknife2 incidence; Chao1; Chao2). The results depend on the input data (true abundances vs counts vs incidences) and users should be careful in the subsequent interpretation. Preston plots are created using the data from the \code{prestondistr} results in \code{ggplot2} featuring cyan bars for observed species (or equivalent taxonomic group) and orange dots for expected counts. The argument \code{presence.absence} will convert the counts (or abundances) to 1 and 0. This dataset can then directly be used as the input data for functions from packages like \code{vegan} for biodiversity analyses.Beta diversity values are calculated using \code{\link[BAT:beta]{BAT::beta()}} function, which partitions the data using the Podani & Schmera (2011)/Carvalho et al. (2012) approach partitioning the beta diversity into ’species replacement’ and ’richness difference’ components. These results are stored as distance matrices in the output.
\verb{\\emph\{Note on the community matrix\}}: Instead of species counts (or abundances), values in each cell of this matrix are the counts (or abundances) of a specific BIN from a \code{site.cat} site category or a \code{grids.cat}. These counts can be generated at any taxonomic hierarchical level for a single or multiple taxa (This can also be done for 'bin_uri'; the difference being that the numbers in each cell would be the number of times that respective BIN is found at a particular \code{site.cat} or \code{grids.cat}). \code{site.cat} can be any of the \code{geography} fields (Meta data on fields can be checked using the \code{\link[=bold.fields.info]{bold.fields.info()}}). Alternatively, \code{grids.cat} = TRUE will generate grids based on the BIN occurrence data (latitude, longitude) with the size of the grid determined by the user (in sq.m.) with the \code{gridsize} argument. For \code{grids} generation, rows with no latitude and longitude data are removed (even if a corresponding \code{site.cat} information is available) while NULL entries for \code{site.cat} are allowed if they have a latitude and longitude value (This is done because grids are drawn based on the bounding boxes which only use latitude and longitude values).
\verb{\\emph\{Note:\}} Results, including species counts, adapt based on \code{taxon.rank} argument although the output label remains ‘species’ in some instances (\code{preston.res}).
}
\examples{
\dontrun{

# Search for ids
comm.mat.data<-bold.public.search(taxonomy = "Poecilia")

# Fetch the data using the ids
BCDMdata<-bold.fetch(param.data = comm.mat.data,query.param = "processid",param.index = 1,apikey)

# Remove rows which have no species data
data<-BCDMdata[!BCDMdata$species=="",]

#1. Analyze richness data (site.cat='country.ocean')
res.rich<-bold.analyze.diversity(BCDMdata,taxon.rank="species",'country.ocean',richness.res=TRUE)

# Community matrix (BCDM data converted to community matrix)
res.rich$comm.matrix

# richness results
res.rich$comm.matrix

#2. Shannon diversity (taxon.rank = 'species'; site.cat='country.ocean')
res.shannon<-bold.analyze.diversity(BCDMdata,"species",'country.ocean',shannon.res = TRUE)

# Shannon diversity results
res.shannon

#3. Preston plots and results (taxon.rank = 'species'; site.cat='country.ocean')
pres.res<-bold.analyze.diversity(BCDMdata,"species",'country.ocean',preston.res=TRUE)

# Preston plot
pres.res$preston.plot

# Preston plot data
pres.res$preston.res

#4. beta diversity (taxon.rank = 'species'; site.cat='country.ocean'; beta.index = "jaccard")
beta.res<-bold.analyze.diversity(BCDMdata,
"species",
'country.ocean',
beta.res=TRUE,
"jaccard")

#Total diversity
beta.res$total.beta

#Replacement
beta.res$replace

#Richness difference
beta.res$richnessd

}

}
