% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bold.apikey.R
\name{bold.apikey}
\alias{bold.apikey}
\title{Set the BOLD private data API key}
\usage{
bold.apikey(apikey)
}
\arguments{
\item{apikey}{A character string required for authentication and data access.}
}
\value{
Token saved as 'apikey'
}
\description{
Stores the BOLD-provided access token ‘api key’ in a variable, making it available for use in other function within the R session.
}
\details{
\code{bold.apikey} creates a variable called \code{apikey} that stores the access token provided by BOLD. This apikey variable is then used internally by the \code{\link[=bold.fetch]{bold.fetch()}} function, so that the user does not have need to input it again. To set the \code{apikey}, the token must be provided as an input for the function before any other functions are called. The api_key is a UUID v4 hexadecimal string and is valid for one year,  after which it must be renewed.
\emph{Obtaining the API key}: The API key is found in the BOLD 'Workbench'\url{https://bench.boldsystems.org/index.php/Login/page?destination=MAS_Management_UserConsole}. After logging in, navigate to 'Your Name' (located at the top left-hand side of the window) and click 'Edit User Preferences'. You can find the API key in the 'User Data' section. Please note that to have an API key available in the workbench, a user must have uploaded at least 10,000 records to BOLD.
}
\examples{
\dontrun{

#This example below is for documentation only

bold_apikey(‘00000000-0000-0000-0000-000000000000’)

}


}
