% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bold.apikey.R
\name{bold.apikey}
\alias{bold.apikey}
\title{Set the BOLD private data API key}
\usage{
bold.apikey(apikey)
}
\arguments{
\item{apikey}{A character string required for authentication and data access.}
}
\value{
Token saved as 'apikey'
}
\description{
Generates an 'apikey' variable in the R session
}
\details{
\code{bold.apikey} generates a variable called \code{apikey} that stores the access token provided by BOLD. The \code{apikey} variable is then used by the \code{\link[=bold.fetch]{bold.fetch()}} internally so that the user does not have to input it again. The token must be provided as an input for the function before any other functions are used so as to set the apikey. The \code{api_key} is a UUID v4 hexadecimal string obtained upon request from BOLD at \code{support@boldsystems.org} and is valid for one year, requiring renewal thereafter.
}
