% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.connectr.public.R
\name{bold.connectr.public}
\alias{bold.connectr.public}
\title{Fetch publicly available data from the BOLD database}
\usage{
bold.connectr.public(
  taxonomy = NULL,
  geography = NULL,
  bins = NULL,
  ids = NULL,
  datasets = NULL,
  latitude = NULL,
  longitude = NULL,
  shapefile = NULL,
  institutes = NULL,
  identified.by = NULL,
  seq.source = NULL,
  marker = NULL,
  collection.period = NULL,
  basecount = NULL,
  altitude = NULL,
  depth = NULL,
  fields = NULL,
  export = NULL,
  file.type,
  file.path,
  file.name
)
}
\arguments{
\item{taxonomy}{A single character string or a vector of taxonomic names at any hierarchical level. Default value is NULL}

\item{geography}{A single character string or a vector of country/province/state/region/sector/site names/codes. Default value is NULL.}

\item{bins}{A single character string or a vector of BIN ids. Default value is NULL.}

\item{ids}{A single character string or a vector of either processids or sampleids. Default value is NULL.}

\item{datasets}{A single character string or a vector of dataset codes. Default value is NULL.}

\item{latitude}{A number or a vector of two numbers indicating the latitudinal range. Values separated by a comma. Default value is NULL.}

\item{longitude}{A number or a vector of two numbers indicating the longitudinal range. Values separated by a comma. Default value is NULL.}

\item{shapefile}{A file path pointing to a shapefile or name of the shapefile imported in the session. Default value is NULL.}

\item{institutes}{A character string or a vector specifying names of institutes. Default value is NULL.}

\item{identified.by}{A character string or a vector specifying names of people responsible for identifying the organism. Default value is NULL.}

\item{seq.source}{A character string or a vector specifying data portals from where the (sequence) data was mined. Default value is NULL.}

\item{marker}{A character string or a vector specifying  of gene names. Default value is NULL.}

\item{collection.period}{A Date or a vector of two values specifying the collection period range (start, end). Values separated by comma. Default value is NULL.}

\item{basecount}{A number or a vector of two numbers indicating the base pairs number range. Values separated by a comma. Default value is NULL.}

\item{altitude}{A number or a vector of two numbers indicating the altitude range. Values separated by a comma. Default value is NULL.}

\item{depth}{A number or a vector of two numbers indicating the depth range. Values separated by a comma. Default value is NULL.}

\item{fields}{A character string or a vector specifying columns needed in the final dataframe. Default value is NULL.}

\item{export}{A logical value asking if the output should be exported locally}

\item{file.type}{A character value specifying the type of file to be exported. Currently ‘.csv’ and ‘.tsv’ options are available}

\item{file.path}{A character value specifying the folder path where the file should be saved}

\item{file.name}{A character value specifying the name of the exported file.}
}
\value{
A data frame containing all the information related to the query search
}
\description{
Retrieves publicly available data based on taxonomy,geography or ids (processid,sampleid, dataset codes & bin_uri) input
}
\details{
Function to retrieve publicly available data on BOLD. Data can be retrieved by providing either one or a combination of taxonomy, geography, bins and/or ids. There is no limit on the data that can be downloaded but complex combinations of the search parameters can lead to weburl limit exceeding (2048 characters). Downloaded data can then be filtered on either one or a combination of arguments (such as institutes, identifiers, altitude, depth etc). Using the ‘fields’ argument will let the user select any specific columns that need to be in the final data frame instead of the whole data. The default NULL value will result in all data being acquired.
}
