% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.data.summarize.R
\name{bold.data.summarize}
\alias{bold.data.summarize}
\title{Generate a summary of the data downloaded from BOLD}
\usage{
bold.data.summarize(bold.df, cols = NULL)
}
\arguments{
\item{bold.df}{the data.frame retrieved from the \code{bold.fetch} function.}

\item{cols}{A single or multiple character vector specifying the columns for which a data summary is sought. Default value is NULL.}
}
\value{
A list of data frames. Each data frame is a data summary of a specific data type.
}
\description{
The function is used to obtain a detailed summary of the data obtained by \code{bold.fetch} function.
}
\details{
\code{bold.data.summarize} provides summaries for each data type available in the downloaded dataset. The function uses the \code{skimr::skim()} function to generate a list of data frames which are then separated by data type using \code{skimr::partition()} to facilitate streamlined export. The summary includes counts for NULL values, unique values and the proportion of complete cases. The cols argument allows users to select specific fields for summarization; by default, it is set to NULL, meaning all columns are summarized. Summaries are printed to the console and can also be saved. Please note that if the fields argument from bold.fetch has been used to filter for specific columns,  and only those will be summarized by default. For specific details on the skim output, refers to the \code{skimr} package documentation.
}
\examples{
\dontrun{
# Download data
bold_data.ids <- bold.public.search(taxonomy = "Oreochromis")

bold.data <- bold.fetch(bold_data.ids,
query.param = "processid",
param.index = 1,
api_key = apikey)

# Generate summary for specific fields (cols)
test.data.summary <- bold.data.summarize(bold_data,
                                         cols = c("country.ocean", "nuc_basecount", "inst", "elev"))

Character data fields summary
test.data.summary$character

Numerical data fields summary
test.data.summary$numeric

}

}
