% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.data.summarize.R
\name{bold.data.summarize}
\alias{bold.data.summarize}
\title{Generate a summary of the data downloaded from BOLD}
\usage{
bold.data.summarize(bold.df, cols = NULL)
}
\arguments{
\item{bold.df}{the data.frame retrieved from the \code{bold.fetch} function.}

\item{cols}{A single or multiple character vector specifying the columns for which a data summary is sought. Default value is NULL.}
}
\value{
A list of data frames. Each data frame is a data summary of a specific data type.
}
\description{
The function is used to obtain a detailed summary of the data obtained by \code{bold.fetch} function.
}
\details{
\code{bold.data.summarize} provides summaries for each data type available in the downloaded dataset. The function uses the \code{\link[skimr:skim]{skimr::skim()}} function  to generate a list of data frames followed by the \code{\link[skimr:partition]{skimr::partition()}} which separates the summary based on the data type for easy export. The summary includes counts for NULL, unique values along with proportion of complete cases. The \code{cols} argument will select any specific field required.The default NULL will generate a summary of all columns. The output is printed on the console and can be saved as well. Please note that if the \code{fields} argument from \code{bold.fetch}has been used to filter certain columns, summaries of only those columns will be available by default.
}
\examples{
\dontrun{
# Download data
bold_data<-bold.connectr.public(taxonomy = "Oreochromis")

# Generate summary for specific fields (cols)
test.data.summary<-data.summary(bold_data,cols = c("country.ocean","nuc_basecount","inst","elev"))

# Character data fields summary
test.data.summary$character

# Numerical data fields summary
test.data.summary$numeric
}

}
