% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bold.data.summarize.R
\name{bold.data.summarize}
\alias{bold.data.summarize}
\title{Generate a summary of the data downloaded from BOLD}
\usage{
bold.data.summarize(
  bold_df,
  summarize_by = c("fields", "presets", "all_data"),
  columns = NULL,
  presets = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{bold_df}{the data.frame retrieved from the \code{bold.fetch} function.}

\item{summarize_by}{A single character value specifying the type of summary required ("fields","presets","all_data" currently available)}

\item{columns}{A single or multiple character vector specifying the columns for which a data summary is sought. Default value is NULL.}

\item{presets}{A single character vector specifying a preset for which a data summary is sought (Check the \code{details} section for more information). Default value is NULL.}

\item{na.rm}{A logical value specifying whether NA values should be removed from the BCDM dataframe. Default value is FALSE.}
}
\value{
A list of data frames. Each data frame is a data summary of a specific data type.
}
\description{
The function is used to obtain a detailed summary of the data obtained by \code{bold.fetch} function.
}
\details{
\code{bold.data.summarize} provides summaries for each data type available in the downloaded dataset. The function uses the \code{skimr::skim()} function to generate a list of data frames which are then separated by data type using \code{skimr::partition()} to facilitate streamlined export. \code{na.rm}= TRUE removes all NA values. Summaries can be generated by one of three ways, a) using \code{all_data} that considers all data, b) using specific \code{presets}that are defined set of columns representing a specific aspect of the BOLD BCDM data (currently available are: \code{taxonomy},\code{geography},\code{sequences},\code{attributions}, \code{ecology_biogeography} and \code{other_meta_data}) and c) \code{fields} which lets the user select any specific columns. \code{presets} names can be directly typed without quotations. The summary includes counts for NULL values, unique values and the proportion of complete cases. Both \code{presets} and \code{fields} by default, are set to NULL. Summaries are printed to the console and can also be saved. Please note that if the \code{cols} argument from \code{\link[=bold.fetch]{bold.fetch()}} has been used to filter for specific columns,  and only those will be summarized by default if \code{all_data} option is selected. For specific details on the skim output, refers to the \code{skimr} package documentation.
}
\examples{
\dontrun{
bold_data.ids <- bold.public.search(taxonomy = "Oreochromis")

# Fetch the data using the ids.
#1. api_key must be obtained from BOLD support before usage
#2. the function `bold.apikey` should be used to set the apikey in the global env

bold.data <- bold.fetch(get_by = "processid",
                        identifiers = bold_data.ids$processid)

#1. Generate summary for the whole data

test.data.summary <- bold.data.summarize(bold_df=bold.data,
                                         summarize_by = "all_data")

# All summary
test.data.summary

#2. Generate summary for specific fields (cols)
test.data.summary.cols <- bold.data.summarize(bold_df=bold.data,
                                              summarize_by = "fields",
                                              columns = c("country.ocean",
                                                          "nuc_basecount",
                                                          "inst",
                                                          "elev"),
                                             na.rm = F)

# Character data fields summary
test.data.summary.cols$character

# Numerical data fields summary
test.data.summary.cols$numeric

#3. Preset based summary
test.data.summary.preset <- bold.data.summarize(bold_df=bold.data,
                                                summarize_by = "presets",
                                                presets = geography,
                                                na.rm = F)

# All summary with respect to geography related fields in the BCDM data
test.data.summary.preset
}

}
