% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.export.R
\name{bold.export}
\alias{bold.export}
\title{Export files generated by BOLDconnectR}
\usage{
bold.export(
  bold.df,
  export = c("mod.df", "msa.fas", "fas"),
  fas.seq.name.fields = NULL,
  df.export.file.type = NULL,
  export.file.path = NULL,
  export.file.name = NULL
)
}
\arguments{
\item{bold.df}{The data.frame either retrieved from \code{\link[=bold.fetch]{bold.fetch()}},\code{\link[=bold.analyze.align]{bold.analyze.align()}} or a user modified BCDM dataset.}

\item{export}{A character input specifying the type of output required. Should be either of "mod.df","msa.fas" or "fas".}

\item{fas.seq.name.fields}{A single or multiple character vector specifying the column headers which to be should be used to name each sequence in the fasta file. Default is NULL in which case, only the processid is used as a name.}

\item{df.export.file.type}{A character value specifying the type of file to be exported for the modified dataframe. Currently ‘.csv’ and ‘.tsv’ options are available.}

\item{export.file.path}{A character value specifying the folder path where the file should be saved.,}

\item{export.file.name}{A character value specifying the name of the exported file.}
}
\value{
It exports a \code{.fas} or a \code{csv/tsv} file based on the \code{export} argument.
}
\description{
The function is used to export some of the output data generated by BOLDconnectR
}
\details{
\code{bold.export} provides an export option for some of the sequence based outputs obtained by functions from \code{BOLDconnectR}. Sequence information downloaded using the \code{bold.fetch.data} or the aligned sequences obtained using \code{bold.analyze.align} can be exported as a fasta file for third party tool use (\code{export}='fas' or 'msa.fas'). Data downloaded by the \code{bold.fetch} can be directly used to export the unaligned fasta file while the modified dataframe obtained after using the \code{bold.analyze.align} is needed for exporting the multiple sequence alignment. Name for individual sequences in the unaligned fasta file output can be customized by using the \code{fas.seq.name.fields} argument. If more than one field is specified, the name will follow the sequence of the fields given in the vector. The multiple sequence aligned fasta file uses the same name given by the user in the \code{bold.analyze.align} function. In addition, the function also allows user edited data (in taxonomy,geography etc.)  to be exported  as a csv/tsv file while retaining its BCDM format. This functionality is added keeping in the mind the possibility of uploading data to BOLD using the package in the near future.The edits done to the BCDM data could be from any other R packages so long as it maintains the BCDM format.
}
\examples{
\dontrun{
# Download data
bold_data<-bold.connectr.public(taxonomy = "Oreochromis")

# Generate summary for specific fields (cols)
test.data.summary<-data.summary(bold_data,cols = c("country.ocean","nuc_basecount","inst","elev"))

# Character data fields summary
test.data.summary$character

# Numerical data fields summary
test.data.summary$numeric
}

}
