% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bold.export.R
\name{bold.export}
\alias{bold.export}
\title{Export files generated by BOLDconnectR}
\usage{
bold.export(
  bold_df,
  export_type = c("preset_df", "msa", "fas"),
  presets = NULL,
  cols_for_fas_names = NULL,
  export_to
)
}
\arguments{
\item{bold_df}{The data.frame either retrieved from \code{\link[=bold.fetch]{bold.fetch()}},\code{bold.analyze.align} or a user modified BCDM dataset.}

\item{export_type}{A character input specifying the type of output required. Should be either of "preset_df","msa" or "fas".}

\item{presets}{A single character vector specifying a preset for which a data summary is sought (Check the \code{details} section for more information). Default value is NULL.}

\item{cols_for_fas_names}{A single or multiple character vector indicating the column headers that should be used to name each sequence for the unaligned FASTA file. Default is NULL; in this case, only the processid is used as the name.}

\item{export_to}{A character value specifying the data path and the name for the file. Extension should not be provided}
}
\value{
It exports a .fas or a csv/tsv file based on the export argument.
}
\description{
The function is used to export some of the output data generated by BOLDconnectR
}
\details{
\code{bold.export} offers an export option for some of the sequence-based outputs obtained from functions within the \code{BOLDconnectR} package as well as a \code{preset} defined modified BCDM dataframe. Sequence information downloaded using \code{\link[=bold.fetch]{bold.fetch()}} or the aligned sequences obtained using \code{bold.analyze.align} can be exported as a FASTA file for any third party tool (via \code{export_type}=’fas’ or ’msa’). Data fetched by \code{\link[=bold.fetch]{bold.fetch()}} can be directly used to export the unaligned FASTA file, while the modified dataframe from \code{bold.analyze.align} is required for exporting the multiple sequence alignment. \code{presets}are defined set of columns representing a specific aspect of the BOLD BCDM data (currently available are: \code{taxonomy},\code{geography},\code{sequences},\code{attributions}, \code{ecology_biogeography} and \code{other_meta_data})
The name for individual sequences in the unaligned FASTA file output can be customized by using the \code{cols_for_fas_names} argument. If more than one field is specified, the name will follow the sequence of the fields given in the vector. The multiple sequence aligned FASTA file uses the same name provided by the user in the \code{bold.analyze.align} function. Additionally, this function allows for the export of user-edited data (in taxonomy, geography etc.) as a csv/tsv file while retaining its BCDM format. This functionality is developed with the future potential of uploading data to BOLD using the package. Edits to the BCDM data can be made using any other R packages so long as it maintains the BCDM format.
}
\examples{
\dontrun{
# Download the records
data_for_export_ids <- bold.public.search(taxonomy = "Poecilia reticulata")

# Fetching the data using the ids
data_for_export <- bold.fetch(get_by = "processid",
                              identifiers = data_for_export_ids$processid)

#1. Export the BCDM data using 'presets'
# (Using getwd() as the path and trial_export as the name)
bold.export(bold_df=data_for_export,
            export_type = "preset_df",
            presets = taxonomy,
            export_to = paste(getwd(),'/','trial_export',sep=""))

#2. Export multiple sequence alignment
#a. Align the data
# (using processid and bin_uri as fields for sequence names)
# Users need to install and load packages `msa` and `Biostrings` before using bold.analyze.align.
seq_align<-bold.analyze.align(data_for_export,
                              cols_for_seq_names = c("processid","bin_uri"),
                              align_method = "ClustalOmega")

#b. Export the multiple sequence alignment
# Note the input data here is the modified BCDM data (seq_align)
bold.export(bold_df=seq_align,
           export_type = "msa",
            export_to = paste(getwd(),'/','trial_export',sep=""))

#3. Export the fasta file (unaligned)
# Note that input data here is the original BCDM data (data_for_export)
bold.export(bold_df = data_for_export,
            export_type = "fas",
            cols_for_fas_names = c("bin_uri","genus","species"),
            export_to = paste(getwd(),'/','trial_export',sep=""))
}

}
