% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.fields.info.R
\name{bold.fields.info}
\alias{bold.fields.info}
\title{Helper function to obtain the BOLD.fields.data}
\usage{
bold.fields.info(print.output = FALSE)
}
\arguments{
\item{print.output}{Whether the output should be printed in the console. Default is FALSE.}
}
\value{
A data frame containing information on all fields (columns)
}
\description{
Helper function to obtain the BOLD.fields.data
}
\details{
The function downloads the latest field (column) names and its meta data (file type and brief description) which is currently available for download from BOLD
}
