% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.fields.info.R
\name{bold.fields.info}
\alias{bold.fields.info}
\title{Retrieve metadata of the BOLD data fields}
\usage{
bold.fields.info(print.output = FALSE)
}
\arguments{
\item{print.output}{Whether the output should be printed in the console. Default is FALSE.}
}
\value{
A data frame containing information on all fields (columns)
}
\description{
#' @description
This function provides information on the field (column) names and their respective data type
}
\details{
The function downloads the latest field (column) meta data (file type and brief description) which is currently available for download from BOLD.\verb{print,output} = TRUE will print the information in the console.
}
\examples{

bold.field.data<-bold.fields.info()
head(bold.field.data,10)

}
