% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/helper.connect.pvt.4_allparams.fetch.boldfields.info.R
\name{bold.fields.info}
\alias{bold.fields.info}
\title{Helper function to obtain the BOLD.fields.data}
\usage{
bold.fields.info(print.output = FALSE)
}
\arguments{
\item{print.output}{Whether the output should be printed in the console. Default is FALSE.}
}
\description{
Helper function to obtain the BOLD.fields.data
}
