% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold.public.search.R
\name{bold.public.search}
\alias{bold.public.search}
\title{Search publicly available data on the BOLD database}
\usage{
bold.public.search(
  taxonomy = NULL,
  geography = NULL,
  bins = NULL,
  datasets = NULL,
  filt.latitude = NULL,
  filt.longitude = NULL,
  filt.shapefile = NULL,
  filt.institutes = NULL,
  filt.identified.by = NULL,
  filt.seq.source = NULL,
  filt.marker = NULL,
  filt.collection.period = NULL,
  filt.basecount = NULL,
  filt.altitude = NULL,
  filt.depth = NULL
)
}
\arguments{
\item{taxonomy}{A single or multiple character vector specifying the taxonomic names at any hierarchical level. Default value is NULL.}

\item{geography}{A single or multiple character vector specifying any of the country/province/state/region/sector/site names/codes. Default value is NULL.}

\item{bins}{A single or multiple character vector specifying the BIN ids. Default value is NULL.}

\item{datasets}{A single or multiple character vector specifying dataset codes. Default value is NULL.}

\item{filt.latitude}{A single or a vector of two numbers specifying the latitudinal range in decimal degrees. Values should be separated by a comma. Default value is NULL.}

\item{filt.longitude}{A single or a vector of two numbers specifying the longitudinal range in decimal degrees. Values should be separated by a comma. Default value is NULL.}

\item{filt.shapefile}{A file path pointing to a shapefile or name of the shapefile (.shp) imported in the R session. Default value is NULL.}

\item{filt.institutes}{A single or multiple character vector specifying names of institutes. Default value is NULL.}

\item{filt.identified.by}{A single or multiple character vector specifying names of people responsible for identifying the organism. Default value is NULL.}

\item{filt.seq.source}{A single or multiple character vector specifying the data portals from where the (sequence) data was mined. Default value is NULL.}

\item{filt.marker}{A single or multiple character vector specifying  of gene names. Default value is NULL.}

\item{filt.collection.period}{A single or a vector of two date values specifying the collection period range (start, end). Values should be separated by a comma. Default value is NULL.}

\item{filt.basecount}{A single or a vector of two numbers specifying range of basepairs number. Values should be separated by a comma. Default value is NULL.}

\item{filt.altitude}{A single or a vector of two numbers specifying the altitude range in meters. Values should be separated by a comma. Default value is NULL.}

\item{filt.depth}{A single or a vector of two numbers specifying the depth range. Values should be separated by a comma. Default value is NULL.}
}
\value{
A data frame containing all the processids and sampleids related to the query search.
}
\description{
Retrieves record ids for publicly available data based on taxonomy,geography or ids (dataset codes & bin_uri) search.
}
\details{
\code{bold.public.search} searches publicly available data on BOLD and be retrieves the associated proccessids and sampleids, data for which can then be retrieved by t\code{bold.fetch}. Search parameters can either be one or a combination of taxonomy, geography, bin uri or dataset codes. There is no limit on the id data that can be downloaded but complex combinations of the search parameters can lead to web-url character length exceeding the predetermined limit (2048 characters). Single parameter searches are exempt from this limit. Multi-parameter searches (taxonomy + geography + bins/datasets; Please see the example: Taxonomy + Geography + BIN id) to obtain a filtered result should be logical otherwise output obtained might either be empty or not correct. Downloaded ids can then be filtered further on either one or a combination of arguments (such as institutes, identifiers, altitude, depth etc.). The \code{filt} or filter parameter argument names must be written explicitly (Ex. \code{filt.institutes} = 'CBG' instead of just 'CBG') to avoid any errors.
}
\examples{

\dontrun{
# Taxonomy
bold.data<-bold.connectr.public(taxonomy = "Panthera leo")
head(bold.data,10)

# Taxonomy and Geography
bold.data.taxo.geo<-bold.connectr.public(taxonomy = "Panthera uncia",geography = "India")
head(bold.data.taxo_geo,10)

# Taxonomy, Geography and BINs
bold.data.taxo.geo.bin<-bold.connectr.public("Panthera leo", "India",bins=c("BOLD:AAD6819"))
}

}
