% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bold.public.search.R
\name{bold.public.search}
\alias{bold.public.search}
\title{Search publicly available data on the BOLD database}
\usage{
bold.public.search(
  taxonomy = NULL,
  geography = NULL,
  bins = NULL,
  dataset_codes = NULL,
  project_codes = NULL
)
}
\arguments{
\item{taxonomy}{A single or multiple character vector specifying the taxonomic names at any hierarchical level. Default value is NULL.}

\item{geography}{A single or multiple character vector specifying any of the country/province/state/region/sector/site names/codes. Default value is NULL.}

\item{bins}{A single or multiple character vector specifying the BIN ids. Default value is NULL.}

\item{dataset_codes}{A single or multiple character vector specifying the dataset codes. Default value is NULL.}

\item{project_codes}{A single or multiple character vector specifying the project codes. Default value is NULL.}
}
\value{
A data frame containing all the processids and sampleids related to the query search.
}
\description{
Retrieves record ids for publicly available data based on taxonomy, geography, bin_uris or datasets/project codes search.
}
\details{
\code{bold.public.search} searches publicly available data on BOLD, retrieving associated proccessids and sampleids, which can then be accessed using \code{bold.fetch}. Search parameters can include one or a combination of taxonomy, geography, bin uris, dataset or project codes. While there is no limit on the amount of ID data that can be downloaded, complex combinations of the search parameters may exceed the predetermined weburl character length (2048 characters). Searches using a single parameter are not subject to this limit. For multiparameter searches (e.g. taxonomy + geography + bins; see the example: Taxonomy + Geography + BIN id), it’s crucial that the parameters are logically combined to ensure accurate and non-empty results.
}
\examples{

#Taxonomy
bold.data <- bold.public.search(taxonomy = "Panthera leo")

#Result
head(bold.data,10)

#Taxonomy and Geography
bold.data.taxo.geo <- bold.public.search(taxonomy = "Panthera uncia",
geography = "India")

#Result
head(bold.data.taxo.geo,10)

#Taxonomy, Geography and BINs
bold.data.taxo.geo.bin <- bold.public.search("Panthera leo",
geography = "India",
bins=c("BOLD:AAD6819"))

#Result
bold.data.taxo.geo.bin

}
