% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOLDconnectR-package.R
\docType{package}
\name{BOLDconnectR-package}
\alias{BOLDconnectR}
\alias{BOLDconnectR-package}
\title{BOLDconnectR: Retrieve, Transform and Analyze the Barcode of Life Data Systems Data in R}
\description{
Facilitate retrieval, transformation and analysis of the data from the Barcode of Life Data Systems (BOLD) database. It allows both public and private user data to be easily downloaded easily into the R environment using a variety of inputs such as: processids, sampleids, taxonomy, geography etc. The package provides data transformation functions to convert data into formats compatible with other R-packages and third party tools, as well as analysis functions for biodiversity, clustering and mapping. Please note that this package currently requires 'msa' and 'Biostrings' to be installed beforehand.
}
\author{
\strong{Maintainer}: Sameer M. Padhye \email{spadhye@uoguelph.ca}

Authors:
\itemize{
  \item Liliana Ballesteros-Mejia \email{cballest@uoguelph.ca}
  \item Sujeevan Ratnasingham \email{sratnasi@uoguelph.ca}
}

}
\keyword{internal}
