% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.connect.pvt.fetch.processids.R
\name{fetch.bold.processid}
\alias{fetch.bold.processid}
\title{bold.fetch.data processid}
\usage{
fetch.bold.processid(data.input, param.index, api_key)
}
\arguments{
\item{data.input}{input data}

\item{param.index}{the column number in the data which has bin_uri data}

\item{api_key}{API key required to fetch the data}
}
\description{
bold.fetch.data processid
}
