% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.msa.gen_ape_file.R
\name{gen.msa.res}
\alias{gen.msa.res}
\title{Helper functions for sequence based data}
\usage{
gen.msa.res(df, alignmethod, ...)
}
\description{
Helper functions for sequence based data
}
\keyword{internal}
