% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.connect.pvt.GETsingleobs.proc_sampid.R
\name{get_data_processid}
\alias{get_data_processid}
\title{GET_data helper functions for single records of processids and sampleids}
\usage{
get_data_processid(input_data, param.index, api.key)
}
\arguments{
\item{input_data}{input data}

\item{param.index}{the column number in the data which has bin_uri data}

\item{api.key}{API key required to fetch the data}
}
\description{
GET_data helper functions for single records of processids and sampleids
}
\keyword{internal}
