% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.fetch.data.allparams.create.id.files.R
\name{id.files}
\alias{id.files}
\title{Helper Function id.files}
\usage{
id.files(ids)
}
\arguments{
\item{ids}{Character string which are converted to temp file/s which then are used by the POST function}
}
\description{
Helper Function id.files
}
\keyword{internal}
