% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/helper.connect.pvt.4_all.params.postapi.res_fetch.R
\name{post.api.res.fetch}
\alias{post.api.res.fetch}
\title{Helper function to fetch the data using POST API}
\usage{
post.api.res.fetch(base.url, query.params, api.key, temp.file)
}
\arguments{
\item{base.url}{base url for POST API}

\item{query.params}{the query parameters for the POST API}

\item{api.key}{the API key needed to access the data}

\item{temp.file}{the tempfile on which data is written for the POST API}
}
\description{
Helper function to fetch the data using POST API
}
