% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.fetch.json_data_download_&_json2df.R
\name{post.api.res.fetch}
\alias{post.api.res.fetch}
\title{Helper Functions: Retrieve the data using POST API and convert it into a data frame}
\usage{
post.api.res.fetch(base.url, query.params, temp.file)
}
\description{
Helper Functions: Retrieve the data using POST API and convert it into a data frame
}
\keyword{internal}
