% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.biodiversity_profile_fun.R
\name{richness_profile}
\alias{richness_profile}
\title{Helper functions for generating a diversity profile}
\usage{
richness_profile(df)
}
\description{
Helper functions for generating a diversity profile
}
\keyword{internal}
