% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.data.R
\docType{data}
\name{test.data}
\alias{test.data}
\title{A data frame with 2101 processids and sampleids}
\format{
\subsection{}{

A data frame with 2205 rows and 2 columns:
\describe{
\item{processid}{Character vector of processids}
\item{sampleid}{Character vector of sampleids corresponding to the processids}
}
}
}
\usage{
test.data
}
\description{
A data frame containing proccessids and sampleids for data retrieval from BOLD database
}
\keyword{datasets}
