% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.data.R
\docType{data}
\name{test.data}
\alias{test.data}
\title{Canadian spider data by Blagoev et al.(2015)}
\format{
A data frame with 1336 rows and 2 columns:
\describe{
\item{processid}{Character vector of processids}
\item{sampleid}{Character vector of sampleids corresponding to the processids}
}
}
\source{
\url{https://onlinelibrary.wiley.com/doi/full/10.1111/1755-0998.12444}
}
\usage{
test.data
}
\description{
The test data comprises 1,336 process and sample IDs from the Salticidae (Arthropoda:Arachnida:Araneae) family, sourced from Canadian spider data published by Blagoev et al. (2015). This publication includes a DNA barcode reference library encompassing 1,018 species of Canadian spiders.
}
\references{
Blagoev, G. A., Dewaard, J. R., Ratnasingham, S., Dewaard, S. L., Lu, L., Robertson, J., ... & Hebert, P. D. (2016). Untangling taxonomy: a DNA barcode reference library for C anadian spiders. Molecular Ecology Resources, 16(1), 325-341.
}
\keyword{datasets}
