% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.geo.R
\name{visualize.geo}
\alias{visualize.geo}
\title{Visualize organism occurrence data on maps}
\usage{
visualize.geo(
  bold.df,
  country = NULL,
  bbox = NULL,
  export = FALSE,
  file.path = NULL,
  file.name = NULL,
  file.type = NULL
)
}
\arguments{
\item{bold.df}{the data.frame retrieved from \code{\link[=bold.connectr]{bold.connectr()}} or \code{\link[=bold.connectr.public]{bold.connectr.public()}}}

\item{country}{A single or multiple character vector of country names. Default is NULL}

\item{bbox}{A numeric vector specifying the min,max values of the latitude and longitude. Default is NULL}

\item{export}{A logical value asking if the output should be exported locally}

\item{file.path}{A character value specifying the folder path where the file should be saved}

\item{file.name}{A character value specifying the name of the exported file.}

\item{file.type}{A character value specifying the type of file to be exported. Currently ‘.csv’ and ‘.tsv’ options are available}
}
\value{
An 'output' list containing:
\itemize{
\item geo.df = A  simple features (sf) ‘data.frame’ containing the geographic data
\item map_plot = A visualization of the occurrences
}
}
\description{
This function creates basic maps of organism occurrence data on different scales
}
\details{
\code{visualize.geo} extracts out the geographic information from the \code{\link[=bold.connectr]{bold.connectr()}} or \code{\link[=bold.connectr.public]{bold.connectr.public()}} output. Data points having NA values for either latitude or longitude or both are removed. Latitude and longitude values are in ‘decimal degrees’ format. Default view includes data mapped on a world shape file downloaded using the \code{\link[rnaturalearth:ne_countries]{rnaturalearth::ne_countries()}} at a 110 scale. If the ‘country’ is specified (single or multiple values), the function will specifically plot the occurrences on the specified country. Alternatively, a bounding box can be defined for a specific region to be visualized. If export = TRUE, an image file will be saved based on the type (jpg, tiff), and the file path. The function also provides a \link{sf} data frame of the GIS data which can be used for any other application/s.
}
\examples{

#Download data
geo.data<-bold.connectr.public(taxonomy = "Musca domestica")

geo.viz<-visualize.geo(geo.data,export = FALSE)

#The [sf] dataframe of the downloaded data
geo.viz$geo.df

# Visualization
geo.viz$plot

}
